TASK-41 [Refactor]: 커스텀 훅, 무한스크롤 로직 분리
